requirements.txt
discord.py==1.7.3
PyNaCl
youtube_dl
requests
mutagen
pydub
environ
tabulate
ffmpeg

#django
django
django-extensions
django-environ
psycopg2

#debbuging/testing
pytest
ipdb